refactor: create select chart wrapper
